+++ packages/tailwindcss/src/index.ts
@@ -498,3 +498,9 @@ function getVersion() {
     return version
   }
 }
+
+export default function postcssPluginWarning() {
+  throw new Error(
+    `It looks like you're trying to use \`tailwindcss\` directly as a PostCSS plugin. The PostCSS plugin has moved to a separate package, so to continue using Tailwind CSS with PostCSS you'll need to install \`@tailwindcss/postcss\` and update your PostCSS configuration.`,
+  )
+}