+++ packages/tailwindcss/tsup.config.ts
@@ -2,18 +2,11 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig([
   {
-    format: ['esm', 'cjs'],
-    minify: true,
-    dts: true,
-    entry: {
-      lib: 'src/index.ts',
-    },
-  },
-  {
     format: ['esm'],
     minify: true,
     dts: true,
     entry: {
+      lib: 'src/index.ts',
       plugin: 'src/plugin.ts',
       colors: 'src/compat/colors.ts',
       'default-theme': 'src/compat/default-theme.ts',
@@ -25,6 +18,7 @@ export default defineConfig([
     dts: true,
     entry: {
       plugin: 'src/plugin.cts',
+      lib: 'src/index.cts',
       colors: 'src/compat/colors.cts',
       'default-theme': 'src/compat/default-theme.cts',
     },